requirements.txt
asgiref==3.8.1
Django==5.1.3
sqlparse==0.5.2
tzdata==2024.2
psycopg2-binary
django-environ==0.11.2
pillow==11.0.0
django-crispy-forms==2.3
crispy-bootstrap5==2024.10
django-allauth[socialaccount]
requests==2.32.3
PyJWT==2.10.0
cryptography==43.0.3
whitenoise==6.8.2
gunicorn==23.0.0